fix(ci): resolve ruff I001 in relationship.py and split llm_client.py
Two CI failures on PR #167:

1. ruff I001 in agents/memory/relationship.py - applied auto-fix (splits aliased imports into per-symbol from-import blocks; verbose but ruff-clean).

2. File size check - agents/llm_client.py was 528 lines (limit 500). Extracted AnthropicProvider and OpenAIProvider into new agents/llm_providers.py module (287 lines); llm_client.py is now 274 lines and re-exports the provider classes to preserve the historical 'from agents.llm_client import AnthropicProvider' import path used by existing tests.
