+++ agents/memory/relationship.py
@@ -18,13 +18,23 @@ import aiosqlite
 from .migrations import _apply_migrations
 from .relationship_mutations import (
     apply_decay as _apply_decay,
+)
+from .relationship_mutations import (
     record_interaction as _record_interaction,
+)
+from .relationship_mutations import (
     seed_trust as _seed_trust,
+)
+from .relationship_mutations import (
     update_trust as _update_trust,
 )
 from .relationship_queries import (
     get_all_relationships as _get_all_relationships,
+)
+from .relationship_queries import (
     get_relationship_summary as _get_relationship_summary,
+)
+from .relationship_queries import (
     get_trust as _get_trust,
 )
 from .relationship_types import (
